requirements.txt
Django==5.2.6
djangorestframework==3.16.1
requests==2.32.3
